+++ gsi-server/src/test/java/de/gematik/idp/gsi/server/services/RequestValidatorTest.java
@@ -324,4 +324,19 @@ class RequestValidatorTest {
         .isInstanceOf(GsiException.class)
         .hasMessageContaining("invalid amr value");
   }
+
+  @Test
+  void test_validateRedirectUri_VALID() {
+    assertDoesNotThrow(
+        () -> RequestValidator.validateRedirectUri("https://diga-dashboard.anydiga.com/code/ce"));
+  }
+
+  @Test
+  void test_validateRedirectUri_INVALID() {
+    assertThatThrownBy(
+            () ->
+                RequestValidator.validateRedirectUri(" https://diga-dashboard.anydiga.com/code/ce"))
+        .isInstanceOf(GsiException.class)
+        .hasMessageContaining("Invalid redirect uri");
+  }
 }